Refactor Hillfort view to use MVP
