+++ app/src/main/AndroidManifest.xml
@@ -41,7 +41,7 @@
       android:name=".activities.SettingsActivity"
       android:parentActivityName=".activities.ProfileActivity" />
     <activity
-      android:name=".activities.HillfortActivity"
+      android:name=".views.hillfort.HillfortView"
       android:parentActivityName=".activities.HillfortListActivity" />
 
     <meta-data
